requirements.txt
jinja2>=3.1
requests>=2.31
python-dotenv>=1.0
